Rename: 모델 파일 위치 변경 #134
